chore(hybridgateway): refactor backend reference handling with utility functions
